refactor(reducer-object): Rename ReducerMap type to ReducerObject
