+++ packages/data-point/lib/entity-types/entity-hash/factory.test.js
@@ -24,7 +24,7 @@ describe('factory#parse loose modifiers', () => {
     })
 
     expect(result.compose[0]).toHaveProperty('type', 'mapKeys')
-    expect(result.compose[0].transform.type).toBe('ReducerMap') // TODO add more tests for ReducerMap?
+    expect(result.compose[0].transform.type).toBe('ReducerObject') // TODO add more tests for ReducerObject?
   })
   test('factory#addKeys', () => {
     const result = factory.create({
@@ -34,7 +34,7 @@ describe('factory#parse loose modifiers', () => {
     })
 
     expect(result.compose[0]).toHaveProperty('type', 'addKeys')
-    expect(result.compose[0].transform.type).toBe('ReducerMap') // TODO add more tests for ReducerMap?
+    expect(result.compose[0].transform.type).toBe('ReducerObject') // TODO add more tests for ReducerObject?
   })
   test('factory#omitKeys', () => {
     const result = factory.create({
@@ -128,10 +128,10 @@ describe('factory#parse composed modifiers', () => {
     })
     // console.log(result.compose[0].transform.keyMap)
     expect(result.compose[0]).toHaveProperty('type', 'addKeys')
-    expect(result.compose[0].transform).toHaveProperty('type', 'ReducerMap')
+    expect(result.compose[0].transform).toHaveProperty('type', 'ReducerObject')
     // expect(result.compose[0].transform.keyMap[0]).toHaveProperty('key', 'a')
     expect(result.compose[1]).toHaveProperty('type', 'mapKeys')
-    expect(result.compose[1].transform).toHaveProperty('type', 'ReducerMap')
+    expect(result.compose[1].transform).toHaveProperty('type', 'ReducerObject')
     // expect(result.compose[1].transform.keyMap[0]).toHaveProperty('key', 'a')
     expect(result.compose[2]).toHaveProperty('type', 'omitKeys')
     expect(result.compose[2].transform).toEqual(['a'])
