+++ packages/data-point/lib/reducer-object/factory.test.js
