+++ packages/data-point/lib/reducer-object/index.js
