+++ packages/data-point/lib/reducer/factory.js
@@ -4,7 +4,7 @@ const _ = require('lodash')
 
 const ReducerPath = require('../reducer-path')
 const ReducerFunction = require('../reducer-function')
-const ReducerMap = require('../reducer-map')
+const ReducerObject = require('../reducer-object')
 const ReducerEntity = require('../reducer-entity')
 const util = require('util')
 
@@ -16,10 +16,10 @@ const reducerTypes = [ReducerPath, ReducerFunction, ReducerEntity]
  * @return {reducer}
  */
 function create (createTransform, source) {
-  // ReducerMap requires an extra parameter, so
+  // ReducerObject requires an extra parameter, so
   // it's not included in the reducerTypes array
-  if (ReducerMap.isType(source)) {
-    return ReducerMap.create(createTransform, source)
+  if (ReducerObject.isType(source)) {
+    return ReducerObject.create(createTransform, source)
   }
 
   const reducer = _.find(reducerTypes, r => r.isType(source))
